Refactor namespaces and update service imports
Restructure namespaces and organize imports for services and controllers in the ECommercePlateform application. Update `using` directives for a modular approach, define and modify interfaces for authentication, city, country, product, and state functionalities. Revise repository and service interfaces to include relevant methods and ensure alignment with updated paths. Update Angular components and test files to reflect new service locations, enhancing code organization, maintainability, and clarity.
