+++ ECommercePlateform.API/Controllers/CityController.cs
@@ -1,5 +1,5 @@
 ﻿using ECommercePlateform.Application.DTOs;
-using ECommercePlateform.Application.Interfaces;
+using ECommercePlateform.Application.Interfaces.ICity;
 using ECommercePlateform.Domain.Exceptions;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
