+++ ECommercePlateform.Application/Interfaces/IProduct/IProductRepository.cs
@@ -1,6 +1,7 @@
+using ECommercePlateform.Application.Interfaces.IGeneral;
 using ECommercePlateform.Domain.Entities;
 
-namespace ECommercePlateform.Application.Interfaces
+namespace ECommercePlateform.Application.Interfaces.IProduct
 {
     public interface IProductRepository : IGenericRepository<Product>
     {
