+++ ECommercePlateform.Application/Interfaces/IState/IStateService.cs
@@ -1,6 +1,6 @@
 ﻿using ECommercePlateform.Application.DTOs;
 
-namespace ECommercePlateform.Application.Interfaces
+namespace ECommercePlateform.Application.Interfaces.IState
 {
     public interface IStateService
     {
