+++ ECommercePlateform.Application/Interfaces/IUserAuth/IAuthService.cs
@@ -1,6 +1,6 @@
 using ECommercePlateform.Application.DTOs;
 
-namespace ECommercePlateform.Application.Interfaces
+namespace ECommercePlateform.Application.Interfaces.IAuth
 {
     public interface IAuthService
     {
