+++ ECommercePlateform.Infrastructure/Repositories/CityRepository.cs
@@ -1,4 +1,4 @@
-﻿using ECommercePlateform.Application.Interfaces;
+﻿using ECommercePlateform.Application.Interfaces.ICity;
 using ECommercePlateform.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
