+++ ECommercePlateform.Infrastructure/Repositories/UserRepository.cs
@@ -1,4 +1,4 @@
-using ECommercePlateform.Application.Interfaces;
+using ECommercePlateform.Application.Interfaces.IAuth;
 using ECommercePlateform.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
