+++ ecommerceplateform.client/src/app/services/country/country.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Country } from '../models/country.model';
-import { environment } from '../../environments/environment';
+import { Country } from '../../models/country.model';
+import { environment } from '../../../environments/environment';
 
 @Injectable({
   providedIn: 'root'
