+++ ecommerceplateform.client/src/app/services/general/favicon.service.ts
