+++ ecommerceplateform.client/src/app/services/no-white-space.service.spec.ts
@@ -1,16 +1,16 @@
-import { TestBed } from '@angular/core/testing';
+//import { TestBed } from '@angular/core/testing';
 
-import { NoWhiteSpaceService } from './no-white-space.service';
+//import { NoWhiteSpaceService } from './no-white-space.service';
 
-describe('NoWhiteSpaceService', () => {
-  let service: NoWhiteSpaceService;
+//describe('NoWhiteSpaceService', () => {
+//  let service: NoWhiteSpaceService;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(NoWhiteSpaceService);
-  });
+//  beforeEach(() => {
+//    TestBed.configureTestingModule({});
+//    service = TestBed.inject(NoWhiteSpaceService);
+//  });
 
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
+//  it('should be created', () => {
+//    expect(service).toBeTruthy();
+//  });
+//});
