+++ ecommerceplateform.client/src/app/services/no-white-space.service.ts
@@ -1,17 +1,17 @@
-import { Injectable } from '@angular/core';
-import { ValidatorFn, AbstractControl } from '@angular/forms';
+//import { Injectable } from '@angular/core';
+//import { ValidatorFn, AbstractControl } from '@angular/forms';
 
-@Injectable({
-  providedIn: 'root'
-})
-export class NoWhiteSpaceService {
+//@Injectable({
+//  providedIn: 'root'
+//})
+//export class NoWhiteSpaceService {
 
-  constructor() { }
+//  constructor() { }
 
-  noWhitespaceValidator(): ValidatorFn {
-  return (control: AbstractControl): { [key: string]: any } | null => {
-    const isWhitespace = (control.value || '').trim().length === 0;
-    return isWhitespace ? { whitespace: true } : null;
-  };
-}
-}
+//  noWhitespaceValidator(): ValidatorFn {
+//  return (control: AbstractControl): { [key: string]: any } | null => {
+//    const isWhitespace = (control.value || '').trim().length === 0;
+//    return isWhitespace ? { whitespace: true } : null;
+//  };
+//}
+//}
